+++ src/test/java/truco/unitarias/ManejadorTrucoPartidaDosContraDos.java
@@ -1,4 +1,4 @@
-package truco.pruebasUnitarias;
+package truco.unitarias;
 
 import java.util.Arrays;
 import java.util.LinkedList;
@@ -8,12 +8,13 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import truco.modelo.Caballo;
 import truco.modelo.Carta;
-import truco.modelo.Cinco;
-import truco.modelo.Cuatro;
 import truco.modelo.Dos;
 import truco.modelo.EnMano;
+import truco.modelo.Tres;
+import truco.modelo.Cinco;
+import truco.modelo.Seis;
+import truco.modelo.Caballo;
 import truco.modelo.Equipo;
 import truco.modelo.Jugable;
 import truco.modelo.Jugada;
@@ -22,18 +23,15 @@ import truco.modelo.ManejadorTruco;
 import truco.modelo.Mesa;
 import truco.modelo.Moderador;
 import truco.modelo.Palo;
-import truco.modelo.Rey;
-import truco.modelo.Seis;
 import truco.modelo.SieteDeBasto;
 import truco.modelo.SieteDeEspada;
 import truco.modelo.StrategyRotacionEnRonda;
-import truco.modelo.Tres;
 import truco.modelo.UnoDeBasto;
 import truco.modelo.UnoDeCopa;
 import truco.modelo.UnoDeEspada;
 import truco.modelo.UnoDeOro;
 
-public class ManejadorTrucoPartidaPicaPicaTest {
+public class ManejadorTrucoPartidaDosContraDos {
 
 	private Equipo equipo1;
 	private Equipo equipo2;
@@ -43,27 +41,19 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 	private Jugable jugador2;
 	private Jugable jugador3;
 	private Jugable jugador4;
-	private Jugable jugador5;
-	private Jugable jugador6;
 	private List<Jugable> jugadores;
 	private Carta unoDeEspada;
 	private Carta unoDeBasto;
 	private Carta sieteDeEspada;
 	private Carta tresDeCopa;
-	private Carta tresDeEspada;
-	private Carta dosDeEspada;
 	private Carta dosDeBasto;
 	private Carta dosDeOro;
 	private Carta unoDeOro;
 	private Carta unoDeCopa;
-	private Carta reyDeBasto;
-	private Carta reyDeOro;
 	private Carta caballoDeEspada;
 	private Carta sieteDeBasto;
 	private Carta seisDeCopa;
 	private Carta cincoDeEspada;
-	private Carta cuatroDeOro;
-	private Carta cuatroDeCopa;
 	private ManejadorTruco manejador = new ManejadorTruco();
 	
 	@Before
@@ -72,39 +62,29 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador1 = new Jugador("1");
 		this.jugador2 = new Jugador("2");
 		this.jugador3 = new Jugador("3");
-		this.jugador4 = new Jugador("4");
-		this.jugador5 = new Jugador("5");
-		this.jugador6 = new Jugador("6");
+		this.jugador4 = new Jugador("4");	
 		
 		this.jugadores = new LinkedList<Jugable>();
-		
 		this.jugadores.add(jugador1);
 		this.jugadores.add(jugador2);
 		this.jugadores.add(jugador3);
 		this.jugadores.add(jugador4);
-		this.jugadores.add(jugador5);
-		this.jugadores.add(jugador6);
+
 		
 		unoDeEspada = new UnoDeEspada();
 		unoDeBasto = new UnoDeBasto();
 		sieteDeEspada = new SieteDeEspada();
 		tresDeCopa = new Tres(Palo.COPA);
-		tresDeEspada = new Tres(Palo.ESPADA);
-		dosDeEspada = new Dos(Palo.ESPADA);
 		dosDeBasto = new Dos(Palo.BASTO);
 	    dosDeOro = new Dos(Palo.ORO);
 		unoDeOro = new UnoDeOro();
 		unoDeCopa = new UnoDeCopa();
-		reyDeBasto = new Rey(Palo.BASTO);
-		reyDeOro = new Rey(Palo.ORO);
 		caballoDeEspada = new Caballo(Palo.ESPADA);
 		sieteDeBasto = new SieteDeBasto();
 		seisDeCopa = new Seis(Palo.COPA);
 		cincoDeEspada = new Cinco(Palo.ESPADA);
-		cuatroDeOro = new Cuatro(Palo.ORO);
-		cuatroDeCopa = new Cuatro(Palo.COPA);
 		
-		List<Carta> cartas = Arrays.asList(unoDeEspada, unoDeBasto, sieteDeEspada, tresDeCopa, tresDeEspada, dosDeEspada, dosDeBasto, dosDeOro, unoDeOro, unoDeCopa, reyDeBasto, reyDeOro, caballoDeEspada, sieteDeBasto, seisDeCopa, cincoDeEspada, cuatroDeOro, cuatroDeCopa);
+		List<Carta> cartas = Arrays.asList(unoDeEspada, unoDeBasto, sieteDeEspada, tresDeCopa, dosDeBasto, dosDeOro, unoDeOro, unoDeCopa, caballoDeEspada, sieteDeBasto, seisDeCopa, cincoDeEspada);
 		
 		mesa = new Mesa(jugadores);
 		
@@ -123,15 +103,7 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		jugador4.recibirCarta(tresDeCopa);
 		jugador4.recibirCarta(unoDeCopa);
 		jugador4.recibirCarta(cincoDeEspada);
-		
-		jugador5.recibirCarta(tresDeEspada);
-		jugador5.recibirCarta(reyDeBasto);
-		jugador5.recibirCarta(cuatroDeOro);
-		
-		jugador6.recibirCarta(dosDeEspada);
-		jugador6.recibirCarta(reyDeOro);
-		jugador6.recibirCarta(cuatroDeCopa);
-		
+
 		moderador = new Moderador(mesa);
 		moderador.setRotacionStrategy(new StrategyRotacionEnRonda(jugadores));
 		
@@ -143,25 +115,18 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.setModerador(this.moderador);
 		this.jugador3.setModerador(this.moderador);
 		this.jugador4.setModerador(this.moderador);
-		this.jugador5.setModerador(this.moderador);
-		this.jugador6.setModerador(this.moderador);
 	    equipo1 = new Equipo();
 	    equipo2 = new Equipo();
 		equipo1.agregarIntegrante(this.jugador1);
 		equipo1.agregarIntegrante(this.jugador3);
-		equipo1.agregarIntegrante(this.jugador5);
 		equipo2.agregarIntegrante(this.jugador2);
 		equipo2.agregarIntegrante(this.jugador4);
-		equipo2.agregarIntegrante(this.jugador6);
 		jugador1.setEquipo(equipo1);
 		jugador2.setEquipo(equipo2);
 		jugador3.setEquipo(equipo1);
 		jugador4.setEquipo(equipo2);
-		jugador5.setEquipo(equipo1);
-		jugador6.setEquipo(equipo2);
 		
 		manejador.setJugadoresEnfrentados(jugadores);
-		
 	}
 	
 	@Test
@@ -171,23 +136,18 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarPrimerCarta();
 		this.jugador3.jugarTercerCarta();
 		this.jugador4.jugarTercerCarta();
-		this.jugador5.jugarTercerCarta();
-		this.jugador6.jugarTercerCarta();
 		this.manejador.resolverJugada(new Jugada());
+		
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
 		jugador1.jugarSegundaCarta();
 		jugador2.jugarSegundaCarta();
 		jugador3.jugarPrimerCarta();
 		jugador4.jugarPrimerCarta();
-		jugador5.jugarSegundaCarta();
-		jugador6.jugarSegundaCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
 		Assert.assertTrue(this.manejador.getGanador() == jugador1.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador3.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador5.getEquipo());	
 	}
 	
 	@Test
@@ -197,8 +157,6 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarPrimerCarta();
 		this.jugador3.jugarTercerCarta();
 		this.jugador4.jugarTercerCarta();
-		this.jugador5.jugarTercerCarta();
-		this.jugador6.jugarTercerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
@@ -206,15 +164,10 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		jugador2.jugarSegundaCarta();
 		jugador3.jugarSegundaCarta();
 		jugador4.jugarSegundaCarta();
-		jugador5.jugarSegundaCarta();
-		jugador6.jugarSegundaCarta();
-
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
-		Assert.assertTrue(this.manejador.getGanador() == jugador2.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador4.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador6.getEquipo());
+		Assert.assertTrue(this.manejador.getGanador() == jugador2.getEquipo());		
 	}
 	
 	@Test
@@ -224,32 +177,24 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarTercerCarta();
 		this.jugador3.jugarTercerCarta();
 		this.jugador4.jugarTercerCarta();
-		this.jugador5.jugarTercerCarta();
-		this.jugador6.jugarTercerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
 		this.jugador1.jugarTercerCarta();
 		this.jugador2.jugarPrimerCarta();
-		this.jugador3.jugarPrimerCarta();
-		this.jugador4.jugarPrimerCarta();
-		this.jugador5.jugarSegundaCarta();
-		this.jugador6.jugarSegundaCarta();
+		this.jugador3.jugarSegundaCarta();
+		this.jugador4.jugarSegundaCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
 		jugador1.jugarSegundaCarta();
 		jugador2.jugarSegundaCarta();
-		jugador3.jugarSegundaCarta();
-		jugador4.jugarSegundaCarta();
-		jugador5.jugarPrimerCarta();
-		jugador6.jugarPrimerCarta();
+		this.jugador3.jugarPrimerCarta();
+		this.jugador4.jugarPrimerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
 		Assert.assertTrue(this.manejador.getGanador() == jugador1.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador3.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador5.getEquipo());	
 	}
 	
 	@Test
@@ -259,8 +204,6 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarTercerCarta();
 		this.jugador3.jugarTercerCarta();
 		this.jugador4.jugarTercerCarta();
-		this.jugador5.jugarTercerCarta();
-		this.jugador6.jugarTercerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
@@ -268,25 +211,18 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarPrimerCarta();
 		this.jugador3.jugarPrimerCarta();
 		this.jugador4.jugarSegundaCarta();
-		this.jugador5.jugarPrimerCarta();
-		this.jugador6.jugarSegundaCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
 		jugador1.jugarTercerCarta();
 		jugador2.jugarSegundaCarta();
-		jugador3.jugarSegundaCarta();
-		jugador4.jugarPrimerCarta();
-		jugador5.jugarSegundaCarta();
-		jugador6.jugarPrimerCarta();
+		this.jugador3.jugarSegundaCarta();
+		this.jugador4.jugarPrimerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
 		Assert.assertTrue(this.manejador.getGanador() == jugador2.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador4.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador6.getEquipo());	
 	}
-	
 	
 	@Test
 	public void testEmpatePrimeraGanaSegundaEquipoUno(){
@@ -295,8 +231,6 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarSegundaCarta();
 		this.jugador3.jugarSegundaCarta();
 		this.jugador4.jugarSegundaCarta();
-		this.jugador5.jugarSegundaCarta();
-		this.jugador6.jugarSegundaCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
@@ -304,16 +238,12 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		jugador2.jugarPrimerCarta();
 		jugador3.jugarPrimerCarta();
 		jugador4.jugarPrimerCarta();
-		jugador5.jugarPrimerCarta();
-		jugador6.jugarPrimerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
 		Assert.assertTrue(this.manejador.getGanador() == jugador1.getEquipo());	
 		Assert.assertTrue(this.manejador.getGanador() == jugador3.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador5.getEquipo());	
 	}
-	
 	
 	@Test
 	public void testEmpatePrimeraGanaSegundaEquipoDos(){
@@ -322,8 +252,6 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarSegundaCarta();
 		this.jugador3.jugarSegundaCarta();
 		this.jugador4.jugarSegundaCarta();
-		this.jugador5.jugarSegundaCarta();
-		this.jugador6.jugarSegundaCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
@@ -331,16 +259,12 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		jugador2.jugarPrimerCarta();
 		jugador3.jugarTercerCarta();
 		jugador4.jugarPrimerCarta();
-		jugador5.jugarTercerCarta();
-		jugador6.jugarPrimerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
 		Assert.assertTrue(this.manejador.getGanador() == jugador2.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador4.getEquipo());
-		Assert.assertTrue(this.manejador.getGanador() == jugador6.getEquipo());
+		Assert.assertTrue(this.manejador.getGanador() == jugador4.getEquipo());	
 	}
-	
 	
 	@Test
 	public void testEmpatanTerceraDefinePrimeraMano(){
@@ -349,8 +273,6 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		this.jugador2.jugarTercerCarta();
 		this.jugador3.jugarTercerCarta();
 		this.jugador4.jugarTercerCarta();
-		this.jugador5.jugarTercerCarta();
-		this.jugador6.jugarTercerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
@@ -358,8 +280,6 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		jugador2.jugarPrimerCarta();
 		jugador3.jugarPrimerCarta();
 		jugador4.jugarPrimerCarta();
-		jugador5.jugarPrimerCarta();
-		jugador6.jugarPrimerCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
 		
@@ -367,39 +287,10 @@ public class ManejadorTrucoPartidaPicaPicaTest {
 		jugador2.jugarSegundaCarta();
 		jugador3.jugarSegundaCarta();
 		jugador4.jugarSegundaCarta();
-		jugador5.jugarSegundaCarta();
-		jugador6.jugarSegundaCarta();
 		this.manejador.resolverJugada(new Jugada());
 		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
 				
 		Assert.assertTrue(this.manejador.getGanador() == jugador1.getEquipo());	
 		Assert.assertTrue(this.manejador.getGanador() == jugador3.getEquipo());	
-		Assert.assertTrue(this.manejador.getGanador() == jugador5.getEquipo());	
 	}
-	
-	/*Arranco tests pero con un criterio de rotacion pica-pica*/
-	
-	/*@Test
-	public void testRondaPicaPica(){
-		
-		RotacionStrategy rotacionPicaPica = new StrategyRotacionPicaPica(this.jugadores);
-		moderador.setRotacionStrategy(rotacionPicaPica);
-		manejador.setJugadoresEnfrentados(rotacionPicaPica.getJugadoresEnfrentados());
-		
-		moderador.rondaFinalizada();
-		manejador.setJugadoresEnfrentados(rotacionPicaPica.getJugadoresEnfrentados());
-		
-		jugador2.jugarPrimerCarta();
-		jugador5.jugarPrimerCarta();
-		manejador.resolverJugada(new Jugada());
-		Assert.assertFalse(this.manejador.alguienGanoDosDeTres());
-		
-		jugador2.jugarSegundaCarta();
-		jugador5.jugarSegundaCarta(); //gana el jugador del equipo 2
-		manejador.resolverJugada(new Jugada());
-		Assert.assertTrue(this.manejador.alguienGanoDosDeTres());
-		Assert.assertTrue(manejador.getGanador() == equipo1);
-		
-	} el tema con este test es que, al pedirle los jugadores a enfrentar del rotacion Strategy se me relacionan con aquellos del moderador
-	 por lo tanto me tira un null pointer exception */
 }
