+++ src/test/java/truco/unitarias/ModeradorTest.java
@@ -1,4 +1,4 @@
-package truco.pruebasUnitarias;
+package truco.unitarias;
 
 import java.util.LinkedList;
