+++ src/test/java/truco/unitarias/RotacionPicaPicaTest.java
@@ -1,4 +1,4 @@
-package truco.pruebasUnitarias;
+package truco.unitarias;
 
 import java.util.LinkedList;
